Preliminary Physics Flag Stuff
Implements some (untested) bitflags to allow powerups to disable certain physics-related parts of the vanilla code.
These flags are not intended to be reset on level transitions (with the exception of entering a level from the overworld).

Also includes a flag for powerups to enable/disable "Easy Yoshi Flight" (whether you need to hold the 'B' button or repeatedly tap it in order to take flight).

A new !player_backup_direction ram define is additionally included in case a powerup may need it.